requirements.txt
presidio-analyzer
presidio-anonymizer
pypdf
pandas
matplotlib
jupyter
PyYAML
RapidFuzz
spacy
sentence-transformers
